ml-service/requirements.txt
fastapi
uvicorn
pydantic
huggingface-hub
llama-cpp-python
